% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_peaks.R
\name{annotate_peaks}
\alias{annotate_peaks}
\title{Annotation of peaks}
\usage{
annotate_peaks(regions, merge = TRUE, mergedist = 0, tssdist = 1500,
  species = c("human", "mouse"))
}
\arguments{
\item{regions}{GRanges object of peaks to annotate}

\item{merge}{Boolean indicating whether to merge peaks or not}

\item{mergedist}{Maximum distance apart for peaks to be merged}

\item{tssdist}{Distance from TSS to classify as a proximal or distal peak}

\item{species}{character string specifying the name of the species. Only
\code{'human'}, and \code{'mouse'} are supported at present.  [DEFAULT = human].}
}
\value{
GRanges object of annotated peaks
}
\description{
This function categorises peaks as either promoter, genebody or intergenic
based on the distance from a known transcription start site (TSS) and gene
overlap.  By default, the distance is set to 1500 bp but can be changed with
the 'tssdist' argument.  The function also allows regions that are within a
certain distance of each other to be merged to form a larger region.  The
merge distance is set as default to 0 (i.e. they have to be touching) but may be
changed with the 'mergedist' argument, and will only merge peaks within the
same region (i.e. distal peaks will not be merged with proximal peaks).
}
