% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_index.R
\name{build_index}
\alias{build_index}
\title{Build insex for HISAT2}
\usage{
build_index(species = c("human", "mouse"))
}
\arguments{
\item{species}{character string specifying the name of the species. Only
\code{'human'}, and \code{'mouse'} are supported at present.  [DEFAULT = human].}
}
\description{
\code{build_index} for mapping reads using HISAT2. This actually downloads the
pre-compiled index available on the HISAT2 website:
\url{ftp://ftp.ccb.jhu.edu/pub/infphilo/hisat2/data}
}
