% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_chip.R
\name{cc_chip}
\alias{cc_chip}
\title{Create cross-correlation plots for ChIP-seq data}
\usage{
cc_chip(bam.files, paired = TRUE, threads = NULL)
}
\arguments{
\item{bam.files}{character vector specifying path to pre-filtered bam files. [DEFAULT = NULL]}

\item{paired}{logical, if TRUE then paired-end reads. [DEFAULT = TRUE]}

\item{threads}{an integer value indicating the number of workers to be used. If NULL then one less than the
maximum number of cores will be used. [DEFAULT = NULL].}
}
\description{
Create cross-correlation plots for ChIP-seq data
}
