% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\docType{data}
\name{chain}
\alias{chain}
\title{LiftOver chain file in rda format for conversion from hg38 to hg19}
\format{An object of class \code{Chain} of length 25.}
\source{
\url{http://hgdownload.cse.ucsc.edu/goldenpath/hg38/liftOver/hg38ToHg19.over.chain.gz}
}
\usage{
chain
}
\description{
A liftOver chain file, where for each region in the hg38 genome
the alignments of the best/longest syntenic regions are used to translate
features to the hg19 genome.
}
\keyword{datasets}
