% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_peaks.R
\name{consensus_peaks}
\alias{consensus_peaks}
\title{Consensus peaks}
\usage{
consensus_peaks(peaks, replicates, plot = FALSE)
}
\arguments{
\item{peaks}{A GRangesList object consisting of one GRanges object for each
set of peaks.}

\item{replicates}{The minimum number of replicates in which a peak must be
present for it to count as a consensus peak.}

\item{plot}{Boolean to indicate whether or not to generate diagnostic plots.
Default = FALSE.}
}
\value{
A GRanges object of consensus peaks
}
\description{
Given a GRangesList with peaks for each sample, this function determines
consensus peaks (found in at least N replicates, where N is defined by the
user)
}
\examples{
\dontrun{
conPeaks <- consensus_peaks(peaks = peaks, replicates = 2)
}
}
