% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bw.R
\name{create_bw}
\alias{create_bw}
\title{Create bigwig files}
\usage{
create_bw(metadata)
}
\arguments{
\item{metadata}{data.frame object created by run_chip or run_atac funcitons}
}
\description{
Create bigwig files
}
