% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2_analysis.R
\name{deseq2_analysis}
\alias{deseq2_analysis}
\title{Run DESeq2 analysis on ChIP/ATAC-seq count data}
\usage{
deseq2_analysis(metadata, species)
}
\arguments{
\item{metadata}{list created by \code{sanity_check_chip} and with raw
count data stored in \code{metadata$counts}.  Created as part of
\code{run_chip}}

\item{species}{character string specifying the name of the species. Only
\code{'human'}, and \code{'mouse'} are supported at present.  [DEFAULT = human].}
}
\description{
Run DESeq2 analysis on ChIP/ATAC-seq count data
}
