% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_analysis.R
\name{exploratory_analysis}
\alias{exploratory_analysis}
\title{Exploratory analysis of DESeq2 data}
\usage{
exploratory_analysis(dds, species, metadata)
}
\arguments{
\item{dds}{DESeqDataSet object created by \code{DESeq2::DESeq}}

\item{species}{character string specifying the name of the species. Only
\code{'human'}, and \code{'mouse'} are supported at present.  [DEFAULT = human].}

\item{metadata}{list created by \code{sanity_check_chip} and with raw
count data stored in \code{metadata$counts}.  Created as part of
\code{run_chip}}
}
\description{
Exploratory analysis of DESeq2 data
}
