% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl_atac.R
\name{fl_atac}
\alias{fl_atac}
\title{Create fragment size plots for ATAC-seq data}
\usage{
fl_atac(bam.files, threads = NULL)
}
\arguments{
\item{bam.files}{character vector specifying path to pre-filtered bam files. [DEFAULT = NULL]}

\item{threads}{an integer value indicating the number of workers to be used. If NULL then one less than the
maximum number of cores will be used. [DEFAULT = NULL].}
}
\description{
Create fragment size plots for ATAC-seq data
}
