% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_files.R
\name{peak2Granges}
\alias{peak2Granges}
\alias{Granges2df}
\alias{Granges2saf}
\title{File conversions}
\usage{
peak2Granges(peakfile, broad = FALSE)

Granges2df(gr, ignore.strand = FALSE)

Granges2saf(gr)
}
\arguments{
\item{peakfile}{Path to MACS2 peak file in narrowPeak or broadPeak format}

\item{broad}{Logical specifying whether broad or narrowpeak}

\item{gr}{GRanges object to convert}

\item{ignore.strand}{Logical indicating whether strand information should be
included}
}
\value{
GRanges object of peak regions

A dataframe created from the GRanges object

A SAF file created from the GRanges object
}
\description{
Function to convert a narrowPeak or broadPeak MACS2 file into a
  GRanges object

Function to convert a GRanges object to a dataframe

Function to convert a GRanges object to a SAF
}
\examples{
\dontrun{
peaks <- peak2Granges("peaks/sample.narrowPeak")
}
}
