% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_peaksummary.R
\name{plot_peaksummary}
\alias{plot_peaksummary}
\title{Plots of peak width and distance to TSS}
\usage{
plot_peaksummary(annotation, tssdist = 1500)
}
\arguments{
\item{annotation}{list created by the
\code{annotate_peaks()} function - this contains information regarding the distance
from the nearest TSS.}

\item{tssdist}{Distance cutoff to determine proximal and distal peaks.}
}
\description{
Function to plot histograms showing distribution of peak widths and distances
of peaks from transcriptional start sites.
}
