% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_alignment2.R
\name{qc_alignment2}
\alias{qc_alignment2}
\title{Perform alignment statistics on reads}
\usage{
qc_alignment2(bam.files, threads)
}
\arguments{
\item{bam.files}{character vector specifying path to bam files}

\item{threads}{positive integer specifying the number of cores to use}
}
\description{
Perform alignment statistics on reads
}
