% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_great.R
\name{run_great}
\alias{run_great}
\title{GREAT analysis}
\usage{
run_great(regions = NULL, species = c("human", "mouse"))
}
\arguments{
\item{regions}{Granges object of genomic regions to submit for GREAT analysis}

\item{species}{character string specifying the name of the species. Only
\code{'human'}, and \code{'mouse'} are supported at present.  [DEFAULT = human].}
}
\description{
Wrapper around rGREAT package to perform pathway analysis of genomic regions.
Uses the basal-plus-extension rule used as default on the GREAT website.
}
