% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_bam.R
\name{shift_bam}
\alias{shift_bam}
\title{Shift ATAC-seq reads}
\usage{
shift_bam(bam.file, species = c("human", "mouse"), paired = TRUE)
}
\arguments{
\item{bam.file}{character string specifying path to bam file}

\item{species}{character string specifying the name of the species. Only
\code{'human'}, and \code{'mouse'} are supported at present.  [DEFAULT = human].}

\item{paired}{logical, if \code{TRUE} then paired end reads}
}
\description{
Tn5 transposase ahs been shown to bind as a dimer and inserts two adaptors into
accessible DNA locations separated by 9bp.  This function offsets reads on the
positive strand by +4bp and those on the negative strand by -5bp to account for
this for downstream analysis.
}
